Spot Lights

+++ PointLight.cpp
@@ -1,7 +1,5 @@
 #include "PointLight.h"
 
-
-
 PointLight::PointLight() : Light()
 {
 	position = glm::vec3(0.0f, 0.0f, 0.0f);
@@ -10,10 +8,10 @@ PointLight::PointLight() : Light()
 	exponent = 0.0f;
 }
 
-PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue, 
-						GLfloat aIntensity, GLfloat dIntensity, 
-						GLfloat xPos, GLfloat yPos, GLfloat zPos, 
-						GLfloat con, GLfloat lin, GLfloat exp) : Light(red, green, blue, aIntensity, dIntensity)
+PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
+					   GLfloat aIntensity, GLfloat dIntensity,
+					   GLfloat xPos, GLfloat yPos, GLfloat zPos,
+					   GLfloat con, GLfloat lin, GLfloat exp) : Light(red, green, blue, aIntensity, dIntensity)
 {
 	position = glm::vec3(xPos, yPos, zPos);
 	constant = con;
@@ -21,9 +19,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
 	exponent = exp;
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation, 
-							GLfloat diffuseIntensityLocation, GLfloat positionLocation, 
-							GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+						  GLuint diffuseIntensityLocation, GLuint positionLocation,
+						  GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
 	glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
 	glUniform1f(ambientIntensityLocation, ambientIntensity);
